+++ destructive_contract.go
@@ -169,6 +169,7 @@ func setOperationMetadata(operations []cli.Operation) {
 	setDestructiveOperations(operations)
 	setOperationPathParameters(operations)
 	setResolutionIndex(operations)
+	setMultipartOperations(operations)
 	registerBetaResolutionMetadata()
 	relaxResolvableArgsValidation()
 }